pyproject.toml
[project]
name = "disney-reservation-notificator"
version = "0.1.0"
description = ""
authors = [
    {name = "ReiVindicatio",email = "reivindicatio0321@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "requests (>=2.32.3,<3.0.0)"
]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
